Extract MyMoneyConsentConfirmation to own component
