+++ apps/store/src/features/carDealership/MyMoneyConsent/MyMoneyConsentConfirmation/MyMoneyConsentConfirmationAtom.ts
@@ -1,0 +1,9 @@
+import { createJSONStorage } from 'jotai/utils'
+import { atomWithToggleAndStorage } from '@/utils/atomWithToggleAndStorage'
+
+const storage = createJSONStorage(() => sessionStorage)
+export const myMoneyConsentConfirmationFlagAtom = atomWithToggleAndStorage(
+  'wasMyMoneyConsentShown',
+  false,
+  storage,
+)
