+++ apps/store/src/features/carDealership/TrialExtensionForm/TrialExtensionForm.tsx
@@ -3,12 +3,11 @@ import styled from '@emotion/styled'
 import { useAtom } from 'jotai'
 import { useTranslation } from 'next-i18next'
 import { useMemo, useState } from 'react'
-import { BankIdIcon, Button, CheckIcon, Dialog, Space, Text, theme } from 'ui'
+import { BankIdIcon, Button, CheckIcon, Space, Text, theme } from 'ui'
 import { ProductItemContainer } from '@/components/ProductItem/ProductItemContainer'
 import { TotalAmount } from '@/components/ShopBreakdown/TotalAmount'
 import { SpaceFlex } from '@/components/SpaceFlex/SpaceFlex'
 import { TextWithLink } from '@/components/TextWithLink'
-import useLocalSessionStorage from '@/hooks/useLocalSessionStorage'
 import type {
   PriceIntentCarTrialExtensionFragment,
   ProductOfferFragment,
@@ -22,14 +21,11 @@ import { type TrialContract } from '../carDealership.types'
 import { ExtensionOfferToggle } from '../ExtensionOfferToggle'
 import { MyMoneyConsent } from '../MyMoneyConsent/MyMoneyConsent'
 import { concentAcceptedAtom } from '../MyMoneyConsent/MyMoneyConsentAtom'
+import { MyMoneyConsentConfirmation } from '../MyMoneyConsent/MyMoneyConsentConfirmation/MyMoneyConsentConfirmation'
+import { myMoneyConsentConfirmationFlagAtom } from '../MyMoneyConsent/MyMoneyConsentConfirmation/MyMoneyConsentConfirmationAtom'
 import { PriceBreakdown } from '../PriceBreakdown'
 import { ProductItemContractContainerCar } from '../ProductItemContractContainer'
 import { EditActionButton } from './EditActionButton'
-import {
-  consentDialogContent,
-  consentDialogMessage,
-  consentDialogWindow,
-} from './TrialExtensionForm.css'
 import { useAcceptExtension } from './useAcceptExtension'
 
 export type MyMoneyConsentProps = {
@@ -70,16 +66,11 @@ export const TrialExtensionForm = ({
     )
   })
 
-  const [wasMyMoneyConsentShown, setWasMyMoneyConsentShown] = useLocalSessionStorage(
-    'wasMyMoneyConsentShown',
-    false,
+  const [wasMyMoneyConsentShown, setWasMyMoneyConsentShown] = useAtom(
+    myMoneyConsentConfirmationFlagAtom,
   )
 
   const shouldShowMyMoneyConsentConfirmation = !concentAccepted && !wasMyMoneyConsentShown
-
-  const markMyMoneyConsentConfirmationAsShown = () => {
-    setWasMyMoneyConsentShown(true)
-  }
 
   const selectedOffer = useMemo(
     () => getSelectedOffer(priceIntent, tierLevel),
@@ -105,6 +96,8 @@ export const TrialExtensionForm = ({
 
     setTierLevel(tierLevel)
   }
+
+  const markMyMoneyConsentConfirmationAsShown = () => setWasMyMoneyConsentShown(true)
 
   const handleClickSign = () => {
     datadogRum.addAction('Car dealership | Click Sign')
@@ -180,47 +173,17 @@ export const TrialExtensionForm = ({
 
         <Space y={1}>
           {shouldShowMyMoneyConsentConfirmation ? (
-            <Dialog.Root
-              onOpenChange={(isOpen) => {
-                // Dialog is closed
-                if (!isOpen) {
-                  markMyMoneyConsentConfirmationAsShown()
-                }
-              }}
+            <MyMoneyConsentConfirmation
+              onClose={markMyMoneyConsentConfirmationAsShown}
+              onContinue={handleClickSign}
             >
-              <Dialog.Trigger asChild>
-                <Button loading={loading}>
-                  <SpaceFlex space={0.5} align="center">
-                    <BankIdIcon />
-                    {requirePaymentConnection ? t('SIGN_AND_PAY_BUTTON') : t('SIGN_BUTTON')}
-                  </SpaceFlex>
-                </Button>
-              </Dialog.Trigger>
-              <Dialog.Content centerContent={true} className={consentDialogContent}>
-                <Dialog.Window className={consentDialogWindow}>
-                  <div className={consentDialogMessage}>
-                    <Dialog.Title asChild>
-                      <Text size="md">Vill du gå vidare utan erbjudandet?</Text>
-                    </Dialog.Title>
-                    <Dialog.Description asChild>
-                      <Text size="md" color="textSecondary">
-                        Sambla ger dig förmånliga räntor på ditt existerande billån.
-                      </Text>
-                    </Dialog.Description>
-                  </div>
-
-                  <SpaceFlex space={0.5} direction="vertical">
-                    <Dialog.Close asChild>
-                      <Button>Gå tillbaka</Button>
-                    </Dialog.Close>
-
-                    <Button onClick={handleClickSign} loading={loading} variant="ghost">
-                      Fortsätt utan erbjudande
-                    </Button>
-                  </SpaceFlex>
-                </Dialog.Window>
-              </Dialog.Content>
-            </Dialog.Root>
+              <Button>
+                <SpaceFlex space={0.5} align="center">
+                  <BankIdIcon />
+                  {requirePaymentConnection ? t('SIGN_AND_PAY_BUTTON') : t('SIGN_BUTTON')}
+                </SpaceFlex>
+              </Button>
+            </MyMoneyConsentConfirmation>
           ) : (
             <Button onClick={handleClickSign} loading={loading}>
               <SpaceFlex space={0.5} align="center">
